refactor: Form 상태 관리 구조 개선
